wip, cleaning up serde, separating out files.
